+++ bin/panxapi.py
@@ -17,6 +17,7 @@
 #
 
 from __future__ import print_function
+from datetime import datetime
 import sys
 import os
 import getopt
@@ -99,7 +100,17 @@ def main():
             xapi.keygen(extra_qs=options['ad_hoc'])
             print_status(xapi, action)
             print_response(xapi, options)
-            print('API key:  "%s"' % xapi.api_key)
+            if (options['api_username'] and options['api_password'] and
+                    options['hostname'] and options['tag']):
+                # .panrc
+                d = datetime.now()
+                print('# %s generated: %s' % (os.path.basename(sys.argv[0]),
+                                              d.strftime('%Y/%m/%d %H:%M:%S')))
+                print('hostname%%%s=%s' % (options['tag'],
+                                           options['hostname']))
+                print('api_key%%%s=%s' % (options['tag'], xapi.api_key))
+            else:
+                print('API key:  "%s"' % xapi.api_key)
 
         if options['show']:
             action = 'show'
@@ -238,6 +249,20 @@ def main():
             print_status(xapi, action)
             print_response(xapi, options)
 
+        if options['report'] is not None:
+            action = 'report'
+            if options['ad_hoc'] is not None:
+                extra_qs_used = True
+            vsys = options['vsys'][0] if len(options['vsys']) else None
+            xapi.report(reporttype=options['report'],
+                        reportname=options['name'],
+                        vsys=vsys,
+                        interval=options['interval'],
+                        timeout=options['job_timeout'],
+                        extra_qs=options['ad_hoc'])
+            print_status(xapi, action)
+            print_response(xapi, options)
+
         if options['op'] is not None:
             action = 'op'
             kwargs = {
@@ -319,6 +344,19 @@ def main():
     sys.exit(0)
 
 
+def passwd_prompt():
+    import getpass
+
+    try:
+        x = getpass.getpass('Password: ')
+    except EOFError:
+        return None
+    except KeyboardInterrupt:
+        sys.exit(0)
+
+    return x
+
+
 def parse_opts():
     options = {
         'delete': False,
@@ -340,6 +378,8 @@ def parse_opts():
         'op': None,
         'export': None,
         'log': None,
+        'report': None,
+        'name': None,
         'src': None,
         'dst': None,
         'move': None,
@@ -392,6 +432,7 @@ def parse_opts():
                     'group=', 'merge', 'nlogs=', 'skip=', 'filter=',
                     'interval=', 'timeout=',
                     'stime=', 'pcapid=', 'text',
+                    'report=', 'name=',
                     ]
 
     try:
@@ -450,6 +491,10 @@ def parse_opts():
             options['export'] = arg
         elif opt == '--log':
             options['log'] = arg
+        elif opt == '--report':
+            options['report'] = arg
+        elif opt == '--name':
+            options['name'] = arg
         elif opt == '--src':
             options['src'] = arg
         elif opt == '--dst':
@@ -471,9 +516,8 @@ def parse_opts():
                 (options['api_username'],
                  options['api_password']) = arg.split(':', 1)
             except ValueError:
-                print('Invalid api_username:api_password: "%s"' %
-                      arg, file=sys.stderr)
-                sys.exit(1)
+                options['api_username'] = arg
+                options['api_password'] = passwd_prompt()
         elif opt == '-P':
             options['port'] = arg
         elif opt == '--serial':
@@ -820,6 +864,8 @@ def usage():
     -o cmd                execute operational command
     --export category     export files
     --log log-type        retrieve log files
+    --report report-type  retrieve reports (dynamic|predefined|custom)
+    --name report-name    report name
     --src src             clone source node xpath
                           export source file/path/directory
     --dst dst             move/clone destination node name
@@ -830,8 +876,8 @@ def usage():
     --clone               clone object at xpath, src xpath
     --override element    override template object at xpath
     --vsys vsys           VSYS for dynamic update/partial commit/
-                          operational command
-    -l api_username:api_password
+                          operational command/report
+    -l api_username[:api_password]
     -h hostname
     -P port               URL port number
     --serial number       serial number for Panorama redirection/
@@ -841,8 +887,8 @@ def usage():
     --nlogs num           retrieve num logs
     --skip num            skip num logs
     --filter filter       log selection filter
-    --interval seconds    log/commit job query interval
-    --timeout seconds     log/commit job query timeout
+    --interval seconds    log/commit/report job query interval
+    --timeout seconds     log/commit/report job query timeout
     --stime time          search time for threat-pcap
     --pcapid id           threat-pcap ID
     -K api_key